+++ components/script/dom/raredata.rs
@@ -32,6 +32,8 @@ pub(crate) struct NodeRareData {
     pub(crate) mutation_observers: Vec<RegisteredObserver>,
     /// Lazily-generated Unique Id for this node.
     pub(crate) unique_id: Option<UniqueId>,
+
+    pub(crate) slottable_data: SlottableData,
 }
 
 #[derive(Default, JSTraceable, MallocSizeOf)]
@@ -39,8 +41,6 @@ pub(crate) struct NodeRareData {
 pub(crate) struct ElementRareData {
     /// <https://dom.spec.whatwg.org/#dom-element-shadowroot>
     /// The ShadowRoot this element is host of.
-    /// XXX This is currently not exposed to web content. Only for
-    ///     internal use.
     pub(crate) shadow_root: Option<Dom<ShadowRoot>>,
     /// <https://html.spec.whatwg.org/multipage/#custom-element-reaction-queue>
     pub(crate) custom_element_reaction_queue: Vec<CustomElementReaction>,
@@ -58,6 +58,4 @@ pub(crate) struct ElementRareData {
     pub(crate) client_rect: Option<LayoutValue<Rect<i32>>>,
     /// <https://html.spec.whatwg.org/multipage#elementinternals>
     pub(crate) element_internals: Option<Dom<ElementInternals>>,
-
-    pub(crate) slottable_data: SlottableData,
 }
