+++ components/script/layout_dom/element.rs
@@ -140,11 +140,6 @@ impl<'dom> ServoLayoutElement<'dom> {
             Some(node) => matches!(node.script_type_id(), NodeTypeId::Document(_)),
         }
     }
-
-    fn assigned_slot(&self) -> Option<Self> {
-        let slot = self.element.get_assigned_slot()?;
-        Some(Self::from_layout_js(slot.upcast()))
-    }
 }
 
 pub enum DOMDescendantIterator<E>
@@ -204,11 +199,7 @@ impl<'dom> style::dom::TElement for ServoLayoutElement<'dom> {
     }
 
     fn traversal_parent(&self) -> Option<Self> {
-        if let Some(assigned_slot) = self.assigned_slot() {
-            Some(assigned_slot)
-        } else {
-            self.as_node().traversal_parent()
-        }
+        self.as_node().traversal_parent()
     }
 
     fn is_html_element(&self) -> bool {
@@ -752,7 +743,7 @@ impl<'dom> ::selectors::Element for ServoLayoutElement<'dom> {
 
     #[allow(unsafe_code)]
     fn assigned_slot(&self) -> Option<Self> {
-        self.assigned_slot()
+        self.as_node().assigned_slot()
     }
 
     fn is_html_element_in_html_document(&self) -> bool {
